backend/train_model.py
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.svm import SVC
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
import joblib

# This is a placeholder for actual model training
# In a real application, you would have a dataset of text labeled with emotions

def train_text_emotion_model():
    # Sample data - replace with real dataset
    data = {
        'text': [
            "I'm feeling so happy today!",
            "This makes me really sad",
            "I'm so angry about this situation",
            "I feel calm and peaceful",
            "I have so much energy right now",
            "I'm in love with this song",
            "What a wonderful day",
            "I'm devastated by the news",
            "This is so frustrating",
            "I'm completely relaxed",
            "Let's party all night",
            "You're the love of my life"
        ],
        'emotion': [
            'happy', 'sad', 'angry', 'calm', 'energetic', 'romantic',
            'happy', 'sad', 'angry', 'calm', 'energetic', 'romantic'
        ]
    }
    
    df = pd.DataFrame(data)
    
    # Vectorize text
    vectorizer = TfidfVectorizer()
    X = vectorizer.fit_transform(df['text'])
    y = df['emotion']
    
    # Split data
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
    
    # Train model
    model = SVC(kernel='linear', probability=True)
    model.fit(X_train, y_train)
    
    # Evaluate
    y_pred = model.predict(X_test)
    print(classification_report(y_test, y_pred))
    
    # Save model and vectorizer
    joblib.dump(model, 'models/text_emotion_model.joblib')
    joblib.dump(vectorizer, 'models/tfidf_vectorizer.joblib')

def train_history_emotion_model():
    # Sample data - replace with real dataset of audio features labeled with emotions
    # This would require a dataset of tracks with audio features and associated emotions
    
    # Placeholder implementation
    print("This would train a model to predict emotion from audio features")
    
    # Save placeholder model
    joblib.dump("placeholder", 'models/history_emotion_model.joblib')

if __name__ == "__main__":
    train_text_emotion_model()
    train_history_emotion_model()