backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
requests==2.31.0
numpy==2.1.3
scikit-learn== 1.6.1
joblib==1.2.0
transformers==4.30.2
torch==2.7.0
pydantic==1.10.7